PTHREADS_ZG(resources) destroyed too early
It was destroyed before destroying EG(symbol_table): resources marked for keeping would be destroyed if stored in local variables inside ::run() and not destroyed in user-land before the end of the process.

PTHREADS_ZG(resources) is now marked as persistent and destroyed in PHP_MSHUTDOWN_FUNCTION instead of PHP_RSHUTDOWN_FUNCTION. Since PTHREADS_ZG(resources) is necessarily left empty at the end of a request, there is no memory waste.