Remove superfluous inheritance preparations
The parent classes will already be prepared and set in the first loop by the
recursive call to pthreads_prepared_entry in pthreads_copy_entry.

Given this, and the fact that the resolve hash table stores the old class
as the hash, the 'prepared' variable will always result in NULL, making the
code unused.