Refactor unbounded anonymous class copying logic
Previously, there was duplicate copying around some parts of the ce.
This has now been properly cleaned up, where any ce copying logic that
is available to all classes is in pthreads_copy_entry, and any other
ce information that is not available straight away to unbounded
anonymous classes is kept in pthreads_complete_entry to be copied later.