Fixed unnecessary method signature BC breaks in PHP 7.1+ In PHP 7.1+, these resulted in the signatures becoming `?typehint`, which raises warnings when upgrading old code which overrides these methods. As it turns out, these parameters are not supposed to accept null at all (see the arginfo). These nullable flags are not necessary; it is sufficient to declare that only X parameters are needed.
This will break code written for pthreads 3.1.7dev which gave null explicitly to these methods; however the fact that these methods accept null is not documented anywhere and I think it is a bug.
This is backwards compatible with code which overrides these methods since parameter type widening allows this.