Add test for crash when modifying static array after thread start
I haven't been able to find the root cause of this issue yet, but it stems from pthreads' handling of copied static arrays.
When pthreads copies a static array, modifying the array on the parent thread triggers heap corruption and other nasties for reasons I am yet to fully understand.

The test case in the PR demonstrates the bug. It needs to be noted that:
- the crash does not occur unless the `PTHREADS_INHERIT_CLASSES` flag is enabled (hence, it is to do with the way static arrays are being handled)
- the crash occurs in GC when removing members of the array (`Test::$array = [];` works, also `unset(Test::$array[0]);`)
- the crash does not occur if the static array is populated with scalar values or arrays, only objects.